transform: emit error
