+++ test/import/broken.js
@@ -1,0 +1,3 @@
+const sf = require('sheetify')
+
+sf('./not-a-file.css')
